+++ src/dataStructures/stack/compile/calculatePostfixExpression.ts
@@ -1,25 +1,32 @@
 // https://leetcode-cn.com/problems/evaluate-reverse-polish-notation/submissions/
 // 150
-export function evalRPN(tokens: string[]): number {
+export function calculatePostfixExpression(tokens: string[]): number {
   const stack: number[] = [];
+  let num1: number = null;
+  let num2: number = null;
+  let token: string = null;
+  let tmp: number = null;
+
   for (let i = 0; i < tokens.length; i++) {
-    const token = tokens[i];
+    token = tokens[i];
 
     if (token === '+') {
       stack.push(stack.pop() + stack.pop());
     } else if (token === '-') {
-      const num2 = stack.pop();
-      const num1 = stack.pop();
+      num2 = stack.pop();
+      num1 = stack.pop();
       stack.push(num1 - num2);
     } else if (token === '*') {
       stack.push(stack.pop() * stack.pop());
     } else if (token === '/') {
-      const num2 = stack.pop();
-      const num1 = stack.pop();
-      stack.push(num1 / num2 > 0 ? Math.floor(num1 / num2) : Math.ceil(num1 / num2));
+      num2 = stack.pop();
+      num1 = stack.pop();
+      tmp = num1 / num2;
+      stack.push(tmp > 0 ? Math.floor(tmp) : Math.ceil(tmp));
     } else {
       stack.push(parseInt(token, 10));
     }
   }
+
   return stack.pop();
 }